// src/generated/types/MaterialListDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type MaterialListDto = {
  /**
   * @type string
   */
  id: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  description: string
  /**
   * @type string
   */
  attachments: string
  /**
   * @type boolean
   */
  isHidden: boolean
  /**
   * @type string
   */
  courseId: string
  /**
   * @type string, date-time
   */
  created_at: Date
  /**
   * @type string, date-time
   */
  updated_at: Date
  /**
   * @type string
   */
  uploadedBy: string
  /**
   * @type string
   */
  fileSize: string
  /**
   * @type array
   */
  fileUrls: string[]
}